+++ components/approval/backend/src/main/resources/META-INF/services/org.camunda.spin.spi.DataFormatConfigurator
@@ -1,1 +1,1 @@
-io.holunda.polyflow.example.process.approval.config.JacksonDataFormatConfigurator
+io.holunda.polyflow.example.infrastructure.JacksonDataFormatConfigurator
